Start of project
I created and began the index.js file with routes to the account creation and the route to where the account form gets submitted afterward.

I also created some of the pug files for the create account, index, and logged in page as well as the routes file.

Only thing I didn't commit to main branch is data folder, bat file, node_modules, and json package files